+++ examples/animated-glb/src/main.ts
@@ -11,6 +11,7 @@ import {
   startEngine,
   stopAnimation,
 } from "@babylonjs/lite";
+import * as lite from "@babylonjs/lite";
 import { showLiteExplorer } from "../../../src";
 import { demoUrl } from "../../demoUrl";
 
@@ -33,4 +34,4 @@ attachControl(camera, canvas, scene);
 
 await registerScene(scene);
 await startEngine(engine);
-showLiteExplorer({ engine, scene, canvas }, { features: { canvasPicking: true }, userGuideUrl: demoUrl("user-guide/") });
+showLiteExplorer({ engine, scene, canvas, lite }, { features: { canvasPicking: true }, userGuideUrl: demoUrl("user-guide/") });