+++ examples/basic/src/main.ts
@@ -13,6 +13,7 @@ import {
   registerScene,
   startEngine,
 } from "@babylonjs/lite";
+import * as lite from "@babylonjs/lite";
 import { showLiteExplorer } from "../../../src";
 import { demoUrl } from "../../demoUrl";
 
@@ -54,6 +55,6 @@ onBeforeRender(scene, (deltaMs) => {
 await registerScene(scene);
 await startEngine(engine);
 showLiteExplorer(
-  { engine, scene, canvas },
+  { engine, scene, canvas, lite },
   { features: { canvasPicking: true }, userGuideUrl: demoUrl("user-guide/") },
 );